requirements.txt
codegen==1.0
dj-database-url==0.5.0
Django==2.0.3
djangorestframework==3.7.7
framework==0.1.0
future==0.16.0
gpiozero==1.4.1
gunicorn==19.7.1
patterns==0.3
psycopg2==2.7.4
pytz==2018.3
serializers==0.2.4
whitenoise==3.3.1
